Update dependencies versions in Cargo.lock and add support for painting hotspots in TerminalView.rs
The commit updates the versions of dependencies in the Cargo.lock file. It updates the versions of "serde" and "serde_derive" from version 1.0.180 to version 1.0.181.

In TerminalView.rs, the commit adds support for painting hotspots. It introduces changes related to hotspots, including the calculation of hotspot link regions and painting hotspot areas for each line.